Alot of changes, also started working on virtio

+++ kernel/process/thread.h
@@ -50,7 +50,7 @@ typedef struct thread {
     /**
      * the thread id (only unique per process)
      */
-    size_t tid;
+    int tid;
 
     /**
      * the thread state
